switch to `Usages` index to using symbol ids for callee, callee_name

+++ opentaint-ir/opentaint-ir-core/src/main/kotlin/org/opentaint/ir/impl/features/UsagesExtension.kt
@@ -87,7 +87,7 @@ private suspend fun JIRClasspath.findMatches(
             JIRClassOrInterfaceImpl(
                 this,
                 it.source
-            ).declaredMethods.filterIndexed { index, jirMethod -> it.offsets.contains(index) }
+            ).declaredMethods.filterIndexed { index, jirMethod -> it.offsets.contains(index.toByte()) }
         }
     }
